Define auth strategies extension
