+++ Application/EdFi.Ods.AdminApp.Management/ClaimSetEditor/ClaimSetFileImportCommand.cs
@@ -6,6 +6,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using EdFi.Ods.AdminApp.Management.Database.Queries;
+using EdFi.Ods.AdminApp.Management.ClaimSetEditor.Extensions;
 
 namespace EdFi.Ods.AdminApp.Management.ClaimSetEditor
 {
